fix: make aside and header client components, ui fixes
