+++ src/app/messages/u/[username]/page.tsx
@@ -1,10 +1,9 @@
 import Conversation from '@/components/features/messages/conversation/Conversation';
-import ConversationHeader from '@/components/features/messages/conversation/ConversationHeader';
 import ConversationInput from '@/components/features/messages/conversation/ConversationInput';
 import { getAuthSession } from '@/lib/auth';
 import { db } from '@/lib/db';
 import { getUserByUsername } from '@/lib/user';
-import { generateParticipantIds } from '@/lib/utils';
+import { generateParticipantIds } from '@/lib/message';
 
 interface ConversationPageProps {
   params: {
@@ -19,7 +18,10 @@ const ConversationPage = async ({ params }: ConversationPageProps) => {
 
   const user = await getUserByUsername(username);
 
-  const participantIds = generateParticipantIds(session?.user?.id, user?.id);
+  const participantIds = await generateParticipantIds(
+    session?.user?.id,
+    user?.id
+  );
 
   const conversation = await db.conversation.findFirst({
     where: {
@@ -39,8 +41,7 @@ const ConversationPage = async ({ params }: ConversationPageProps) => {
   });
 
   return (
-    <main className='flex-1 justify-between flex flex-col h-[calc(100svh-1rem)] md:h-[calc(100svh-4rem)] relative'>
-      {/* <ConversationHeader user={user} /> */}
+    <main className='flex-1 justify-between flex flex-col h-[calc(100svh-0.8rem)] md:h-[calc(100svh-8rem)] relative'>
       <Conversation conversation={conversation} />
       <ConversationInput authorId={session?.user.id} />
     </main>
