+++ src/lib/pathname.ts
@@ -1,9 +1,9 @@
 import { headers } from 'next/headers';
 
-export function getPathnameFromServer() {
+export function usePathnameOnServer() {
   const headersList = headers();
   const url = headersList.get('referer');
   const urlParts = url?.split('/');
   const pathname = urlParts && urlParts[urlParts.length - 1];
-  return pathname;
+  return { pathname, url };
 }
